#29 fix: interest 도메인을 wish 도메인으로 변경
