--- backend/src/main/java/codesquard/app/domain/interest/InterestRepository.java
@@ -1,6 +1,0 @@
-package codesquard.app.domain.interest;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface InterestRepository extends JpaRepository<Interest, Long> {
-}
