+++ backend/src/main/java/codesquard/app/domain/member/Member.java
@@ -15,7 +15,6 @@ import javax.persistence.Id;
 import javax.persistence.OneToMany;
 
 import codesquard.app.domain.chat.ChatRoom;
-import codesquard.app.domain.interest.Interest;
 import codesquard.app.domain.item.Item;
 import codesquard.app.domain.membertown.MemberTown;
 import lombok.AccessLevel;
@@ -39,9 +38,6 @@ public class Member {
 
 	@OneToMany(cascade = CascadeType.ALL)
 	private List<MemberTown> towns = new ArrayList<>(); // 동네
-
-	@OneToMany(mappedBy = "member", cascade = CascadeType.ALL)
-	private List<Interest> interests = new ArrayList<>(); // 관심하트
 
 	@OneToMany(mappedBy = "member", cascade = CascadeType.ALL)
 	private List<Item> items = new ArrayList<>(); // 회원이 등록한 상품
@@ -71,15 +67,6 @@ public class Member {
 		}
 		if (item != null && !items.contains(item)) {
 			items.add(item);
-		}
-	}
-
-	public void addInterest(Interest interest) {
-		if (interest != null && interest.getMember() != this) {
-			interest.setMember(this);
-		}
-		if (interest != null && !interests.contains(interest)) {
-			interests.add(interest);
 		}
 	}
